requirements.txt
tensorflow==2.12.0
numpy
pillow
matplotlib
opencv-python
scikit-learn
streamlit
